Chapter1/conda_requirements.txt
# packages in environment at /anaconda3/envs/chap1:
#
# Name                    Version                   Build  Channel
_tflow_select             2.2.0                     eigen  
absl-py                   0.7.0                    pypi_0    pypi
appnope                   0.1.0            py36hf537a9a_0  
astor                     0.7.1                    py36_0  
backcall                  0.1.0                    py36_0  
blas                      1.0                         mkl  
bleach                    3.1.0                    py36_0  
c-ares                    1.15.0               h1de35cc_1  
ca-certificates           2019.1.23                     0  
certifi                   2018.11.29               py36_0  
cycler                    0.10.0           py36hfc81398_0  
decorator                 4.3.2                    py36_0  
entrypoints               0.3                      py36_0  
freetype                  2.9.1                hb4e5f40_0  
gast                      0.2.2                    pypi_0    pypi
grpcio                    1.19.0                   pypi_0    pypi
h5py                      2.9.0            py36h3134771_0  
hdf5                      1.10.4               hfa1e0ec_0  
html5lib                  0.999                    py36_0  
icu                       54.1                          0  
intel-openmp              2019.1                      144  
ipykernel                 5.1.0            py36h39e3cac_0  
ipython                   7.3.0            py36h39e3cac_0  
ipython_genutils          0.2.0            py36h241746c_0  
ipywidgets                7.4.2                    py36_0  
jedi                      0.13.3                   py36_0  
jinja2                    2.10                     py36_0  
jsonschema                2.6.0            py36hb385e00_0  
jupyter                   1.0.0                    py36_7  
jupyter_client            5.2.4                    py36_0  
jupyter_console           6.0.0                    py36_0  
jupyter_core              4.4.0                    py36_0  
keras-applications        1.0.7                    pypi_0    pypi
keras-preprocessing       1.0.9                    pypi_0    pypi
kiwisolver                1.0.1            py36h0a44026_0  
libcxx                    4.0.1                hcfea43d_1  
libcxxabi                 4.0.1                hcfea43d_1  
libedit                   3.1.20181209         hb402a30_0  
libffi                    3.2.1                h475c297_4  
libgfortran               3.0.1                h93005f0_2  
libpng                    1.6.36               ha441bb4_0  
libprotobuf               3.6.0                hd9629dc_0  
libsodium                 1.0.16               h3efe00b_0  
markdown                  3.0.1                    py36_0  
markupsafe                1.1.1            py36h1de35cc_0  
matplotlib                3.0.2            py36h54f8f79_0  
mistune                   0.8.4            py36h1de35cc_0  
mkl                       2018.0.3                      1  
mkl_fft                   1.0.6            py36hb8a8100_0  
mkl_random                1.0.1            py36h5d10147_1  
mock                      2.0.0                    pypi_0    pypi
nbconvert                 5.3.1                    py36_0  
nbformat                  4.4.0            py36h827af21_0  
ncurses                   6.1                  h0a44026_1  
notebook                  5.7.4                    py36_0  
numpy                     1.16.2                   pypi_0    pypi
numpy-base                1.15.4           py36h8a80b8c_0  
openssl                   1.1.1b               h1de35cc_1  
pandoc                    2.2.3.2                       0  
pandocfilters             1.4.2                    py36_1  
parso                     0.3.4                    py36_0  
path.py                   10.3.1                   py36_0  
pbr                       5.1.3                    pypi_0    pypi
pexpect                   4.6.0                    py36_0  
pickleshare               0.7.5                    py36_0  
pip                       19.0.3                   pypi_0    pypi
prometheus_client         0.6.0                    py36_0  
prompt_toolkit            2.0.9                    py36_0  
protobuf                  3.7.0                    pypi_0    pypi
ptyprocess                0.6.0                    py36_0  
pygments                  2.3.1                    py36_0  
pyparsing                 2.3.1                    py36_0  
pypng                     0.0.19                   pypi_0    pypi
pyqt                      5.6.0            py36h655552a_6  
python                    3.6.8                haf84260_0  
python-dateutil           2.8.0                    py36_0  
pytz                      2018.9                   py36_0  
pyzmq                     18.0.0           py36h0a44026_0  
qt                        5.6.2                         2  
qtconsole                 4.4.3                    py36_0  
readline                  7.0                  h1de35cc_5  
scipy                     1.1.0            py36h28f7352_1  
send2trash                1.5.0                    py36_0  
setuptools                40.8.0                   py36_0  
simplegeneric             0.8.1                    py36_1  
sip                       4.18.1           py36h0a44026_2  
six                       1.12.0                   py36_0  
sqlite                    3.26.0               ha441bb4_0  
tensorboard               1.13.1                   pypi_0    pypi
tensorflow                1.13.1                   pypi_0    pypi
tensorflow-base           1.11.0          eigen_py36h4f0eeca_0  
tensorflow-estimator      1.13.0                   pypi_0    pypi
termcolor                 1.1.0                    pypi_0    pypi
terminado                 0.8.1                    py36_1  
testpath                  0.4.2                    py36_0  
tk                        8.6.8                ha441bb4_0  
tornado                   5.1.1            py36h1de35cc_0  
traitlets                 4.3.2            py36h65bd3ce_0  
wcwidth                   0.1.7            py36h8c6ec74_0  
webencodings              0.5.1                    py36_1  
werkzeug                  0.14.1                   pypi_0    pypi
wheel                     0.33.1                   py36_0  
widgetsnbextension        3.4.2                    py36_0  
xz                        5.2.4                h1de35cc_4  
zeromq                    4.3.1                h0a44026_3  
zlib                      1.2.11                        0  
